+++ makeJms2.sh
@@ -4,6 +4,12 @@
 #
 # This script creates a copy of the Boot Starter code which is functionally (almost) identical
 # but has the jakarta package names replaced by javax in most places. 
+# 
+# Note that from version 3.2.2, the JMS3 code cannot be simply replicated in this way. Classes
+# that are unique to Spring 3 are now being used. However, this script is being left in here for now
+# as an example; it should still be possible to convert the Spring 3 code to Spring 2, but additional
+# filtering/changes are required. There is a minimal mockup of the SslBundles interface included here
+# that at least permits the code to compile.
 
 curdir=`pwd`
 in="$1"
@@ -35,6 +41,7 @@ do
       grep -v "DeprecatedConfigurationProperty" |\
       sed "s/jakarta/javax/g" |\
       sed "s/ibm.mq.javax/ibm.mq/g" |\
+      sed "s/import org.springframework.boot.ssl.*;//g" |\    
       sed "s/ibm.msg.client.javax/ibm.msg.client/g ">  $out/$f
 done
 
@@ -45,6 +52,29 @@ do
   n=`echo $f | sed "s/.jms2$//g"`
   cp $f $out/$n
 done
+
+# This creates a minimal mockup of the Spring Boot 3 SslBundles interface 
+# in an attempt to allow the JMS3 code to compile as JMS2. 
+# But it's not been seriously tested.
+cat << EOF >$out/src/main/java/com/ibm/mq/spring/boot/SslBundles.java
+package com.ibm.mq.spring.boot;
+import javax.net.ssl.SSLContext;
+class SslBundles {
+   SslBundle getBundle(String s) throws NoSuchSslBundleException{
+     return null;
+   }
+}
+
+class SslBundle {
+   SSLContext createSslContext() {
+     return null;
+   }
+}
+
+class NoSuchSslBundleException extends Exception {
+}
+
+EOF
 
 # One final piece of patching is for the SSLBundles support only available from Spring Boot 3.1
 f="./src/main/resources/META-INF/spring/org.springframework.boot.autoconfigure.AutoConfiguration.imports"