+++ mq-jms-spring-boot-starter/src/main/java/com/ibm/mq/spring/boot/MQConnectionFactoryFactory.java
@@ -1,5 +1,5 @@
 /*
- * Copyright © 2018,2023 IBM Corp. All rights reserved.
+ * Copyright © 2018,2024 IBM Corp. All rights reserved.
  *
  * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file
  * except in compliance with the License. You may obtain a copy of the License at
@@ -16,25 +16,36 @@ package com.ibm.mq.spring.boot;
 
 import java.lang.reflect.Field;
 import java.lang.reflect.InvocationTargetException;
+import java.security.KeyManagementException;
+import java.security.NoSuchAlgorithmException;
 import java.util.Collections;
 import java.util.Hashtable;
 import java.util.List;
 import java.util.Map;
 
-import jakarta.jms.JMSException;
 import javax.naming.Context;
 import javax.naming.NamingException;
 import javax.naming.directory.InitialDirContext;
+import javax.net.ssl.KeyManager;
+import javax.net.ssl.SSLContext;
 import javax.net.ssl.SSLSocketFactory;
+import javax.net.ssl.TrustManager;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
+import org.springframework.boot.ssl.NoSuchSslBundleException;
+import org.springframework.boot.ssl.SslBundle;
+import org.springframework.boot.ssl.SslBundles;
+import org.springframework.boot.ssl.SslManagerBundle;
 
 import com.ibm.mq.jakarta.jms.MQConnectionFactory;
 import com.ibm.msg.client.jakarta.wmq.WMQConstants;
 
+import jakarta.jms.JMSException;
+
 /**
- * Factory to create a {@link MQConnectionFactory} instance from properties defined in {@link MQConfigurationProperties}.
+ * Factory to create a {@link MQConnectionFactory} instance from properties
+ * defined in {@link MQConfigurationProperties}.
  */
 public class MQConnectionFactoryFactory {
 
@@ -42,12 +53,18 @@ public class MQConnectionFactoryFactory {
 
   private final List<MQConnectionFactoryCustomizer> factoryCustomizers;
   private static Logger logger = LoggerFactory.getLogger(MQConnectionFactoryFactory.class);
+  private SslBundles sslBundles;
 
   @SuppressWarnings("unchecked")
-  public MQConnectionFactoryFactory(MQConfigurationProperties properties, List<MQConnectionFactoryCustomizer> factoryCustomizers) {
+  public MQConnectionFactoryFactory(MQConfigurationProperties properties, SslBundles sslBundles,
+      List<MQConnectionFactoryCustomizer> factoryCustomizers) {
     this.properties = properties;
-    this.factoryCustomizers = (List<MQConnectionFactoryCustomizer>) (factoryCustomizers != null ? factoryCustomizers : Collections.emptyList());
+    this.sslBundles = sslBundles;
+    this.factoryCustomizers = (List<MQConnectionFactoryCustomizer>) (factoryCustomizers != null ? factoryCustomizers
+        : Collections.emptyList());
     logger.trace("constructor");
+    // logger.trace("!! Bundles = {} ", (sslBundles == null) ? "null" : (getSSLSocketFactory("ibmmq") != null) ? "Has IBMMQ"
+    // : "No IBMMQ bundle");
   }
 
   @SuppressWarnings("unchecked")
@@ -55,20 +72,23 @@ public class MQConnectionFactoryFactory {
     String err = null;
     T cf = null;
     SSLSocketFactory sf = null;
-       
+
     String jndiProviderUrl = this.properties.getJndi().getProviderUrl();
     String jndiCF = this.properties.getJndi().getProviderContextFactory();
 
     logger.trace("createConnectionFactory for class " + factoryClass.getSimpleName());
-    
-    /* Keystore System properties don't need the CF to be already created */ 
+
+    /* Keystore System properties don't need the CF to be already created */
     String sslBundle = this.properties.getSslBundle();
-    
-    /* From Spring Boot 3.1, we can put sets of SSL configuration items in a bundle */
-    /* The bundle name takes priority over the ibm.mq.jks properties */
-    if (MQConfigurationSslBundles.isSupported() && isNotNullOrEmpty(sslBundle)) {
-      sf = MQConfigurationSslBundles.getSSLSocketFactory(sslBundle);
-    } else {
+
+    /*
+     * From Spring Boot 3.1, we can put sets of SSL configuration items in a bundle
+     * The bundle name takes priority over the ibm.mq.jks properties
+     */
+    if (sslBundles != null && isNotNullOrEmpty(sslBundle)) {
+      sf = getSSLSocketFactory(sslBundle);
+    }
+    else {
       configureTLSStores(this.properties);
     }
 
@@ -86,11 +106,11 @@ public class MQConnectionFactoryFactory {
         else {
           cf = (T) ctx.lookup(cfName);
         }
-        
+
         if (sf != null) {
           cf.setSSLSocketFactory(sf);
         }
-        
+
         // We will not dump the properties as they are not used for most of the configuration.
         // The JNDI config may well have overridden the actual values in any resource files.
         // But we will still allow the customize methods to be used.
@@ -110,21 +130,21 @@ public class MQConnectionFactoryFactory {
         configureConnectionFactory(cf, this.properties);
         customize(cf);
       }
-      catch (JMSException | InstantiationException | IllegalAccessException | IllegalArgumentException | InvocationTargetException | NoSuchMethodException
-          | SecurityException ex) {
+      catch (JMSException | InstantiationException | IllegalAccessException | IllegalArgumentException | InvocationTargetException
+          | NoSuchMethodException | SecurityException ex) {
         logger.trace("createConnectionFactory : exception " + ex.getMessage());
         throw new IllegalStateException("Unable to create MQConnectionFactory" + ((err != null) ? (": " + err) : ""), ex);
       }
     }
-    
+
     return cf;
   }
 
   /*
-  * This method allows someone to create their own CF and then have it configured
-  * using the same MQConfigurationProperties class - which might have been assigned
-  * from a different prefix in the properties file.
-  */
+   * This method allows someone to create their own CF and then have it configured
+   * using the same MQConfigurationProperties class - which might have been assigned
+   * from a different prefix in the properties file.
+   */
   public static void configureConnectionFactory(MQConnectionFactory cf, MQConfigurationProperties props) throws JMSException {
     // Should usually provide a queue manager name but it can be empty, to connect to the
     // default queue manager.
@@ -138,9 +158,10 @@ public class MQConnectionFactoryFactory {
     if (!isNullOrEmpty(qmName)) {
       cf.setStringProperty(WMQConstants.WMQ_QUEUE_MANAGER, qmName);
     }
-    
-    // Use the channel name to decide whether to try to connect locally or as a client. If the queue manager
-    // code has been installed locally, then this connection will try to use native JNI bindings to match.
+
+    // Use the channel name to decide whether to try to connect locally or as a
+    // client. If the queue manager code has been installed locally, then this connection
+    // will try to use native JNI bindings to match.
     String channel = props.getChannel();
     String connName = props.getConnName();
     String ccdtUrl = props.getCcdtUrl();
@@ -170,39 +191,51 @@ public class MQConnectionFactoryFactory {
       if (!isNullOrEmpty(props.getReconnect())) {
         cf.setIntProperty(WMQConstants.WMQ_CLIENT_RECONNECT_OPTIONS, props.getReconnectValue());
       }
-      
+      cf.setIntProperty(WMQConstants.WMQ_CLIENT_RECONNECT_TIMEOUT, props.getReconnectTimeout());
+
       /* Balancing options for Uniform clusters came available from 9.3.4 */
       if (!isNullOrEmpty(props.getBalancingApplicationType())) {
-        cf.setIntProperty(WMQConstants.WMQ_BALANCING_APPLICATION_TYPE,props.getBalancingApplicationTypeValue());
+        cf.setIntProperty(WMQConstants.WMQ_BALANCING_APPLICATION_TYPE, props.getBalancingApplicationTypeValue());
       }
 
       if (!isNullOrEmpty(props.getBalancingOptions())) {
-        cf.setIntProperty(WMQConstants.WMQ_BALANCING_OPTIONS,props.getBalancingOptionsValue());
+        cf.setIntProperty(WMQConstants.WMQ_BALANCING_OPTIONS, props.getBalancingOptionsValue());
       }
 
       if (!isNullOrEmpty(props.getBalancingTimeout())) {
-        cf.setIntProperty(WMQConstants.WMQ_BALANCING_TIMEOUT,props.getBalancingTimeoutValue());
+        cf.setIntProperty(WMQConstants.WMQ_BALANCING_TIMEOUT, props.getBalancingTimeoutValue());
       }
     }
+
     String applicationName = props.getApplicationName();
     if (!isNullOrEmpty(applicationName)) {
       cf.setAppName(applicationName);
     }
 
-    // Setup the authentication. If there is a userid defined, prefer to use the CSP model for
-    // password checking. That is more general than the cf.connect(user,pass) method which has
-    // some restrictions in the MQ client. But it is possible to override the choice via a
+    // Setup the authentication.If a token has been set, then we will use that with an empty
+    // userid, regardless of the configured user property. Otherwise, if
+    // there is a userid defined, prefer to use the CSP model for
+    // password checking.But it is possible to override the choice via a
     // property, for some compatibility requirements. It is possible to have a blank userid
-    // while also setting a password - the queue manager authentication mechanism would then
-    // be responsible for allocating an identity.
-    String u = props.getUser();
-    if (!isNullOrEmpty(u)) {
-      cf.setStringProperty(WMQConstants.USERID, u);
-    }  
-    String p = props.getPassword();
-    if (!isNullOrEmpty(p)) {
-      cf.setStringProperty(WMQConstants.PASSWORD, p);
-      cf.setBooleanProperty(WMQConstants.USER_AUTHENTICATION_MQCSP, props.isUseAuthenticationMQCSP());
+    // while also setting a password. The queue manager authentication mechanism would then
+    // be responsible for allocating an identity - which might come from treating the password
+    // as a token.
+    String token = props.getToken();
+    if (!isNullOrEmpty(token)) {
+      cf.setStringProperty(WMQConstants.PASSWORD, token);
+      cf.setStringProperty(WMQConstants.USERID, "");
+      cf.setBooleanProperty(WMQConstants.USER_AUTHENTICATION_MQCSP, true);
+    }
+    else {
+      String u = props.getUser();
+      if (!isNullOrEmpty(u)) {
+        cf.setStringProperty(WMQConstants.USERID, u);
+      }
+      String p = props.getPassword();
+      if (!isNullOrEmpty(p)) {
+        cf.setStringProperty(WMQConstants.PASSWORD, p);
+        cf.setBooleanProperty(WMQConstants.USER_AUTHENTICATION_MQCSP, props.isUseAuthenticationMQCSP());
+      }
     }
 
     if (!isNullOrEmpty(props.getSslCipherSuite()))
@@ -229,7 +262,7 @@ public class MQConnectionFactoryFactory {
     if (!isNullOrEmpty(props.getTempModel())) {
       cf.setStringProperty(WMQConstants.WMQ_TEMPORARY_MODEL, props.getTempModel());
     }
-    
+
     /*
      * Additional properties that are not in the pre-defined recognised set can be put onto the
      * CF via a map in the external properties definitions. Use the format
@@ -273,7 +306,8 @@ public class MQConnectionFactoryFactory {
 
       try {
         if (v.toUpperCase().startsWith("0X")) {
-          vi = Integer.decode(v); // Could use decode for all values but this reduces the number of strings that might mistakenly match
+          vi = Integer.decode(v); // Could use decode for all values but this reduces the number of strings that
+                                  // might mistakenly match
         }
         else {
           vi = Integer.valueOf(v);
@@ -305,15 +339,15 @@ public class MQConnectionFactoryFactory {
   /*
    * Access to Java keystores can be controlled by system properties. These are usually
    * given with -D options on the command line but we can set them here instead. The Spring properties
-   * that drive these will be "ibm.mq.jks.keyStore=" ...  For historic reasons, we set the com.ibm.ssl versions
-   * as well as the regular javax.net.ssl properties. Deprecated in Spring Boot 3 where we try to use
-   * SSLBundles instead.
+   * that drive these will be "ibm.mq.jks.keyStore=" ... For historic reasons, we set the com.ibm.ssl versions
+   * as well as the regular javax.net.ssl properties. Deprecated in Spring Boot 3
+   * where we try to use SSLBundles instead.
    */
   private static void configureTLSStores(MQConfigurationProperties props) {
-    String prefixes[] = {"javax.net.ssl.","com.ibm.ssl."};
-    
+    String prefixes[] = { "javax.net.ssl.", "com.ibm.ssl." };
+
     logger.trace("configuring TLS Store system properties");
-  
+
     MQConfigurationPropertiesJks jksProperties = props.getJks();
     for (String prefix : prefixes) {
       if (!isNullOrEmpty(jksProperties.getKeyStore())) {
@@ -331,8 +365,8 @@ public class MQConnectionFactoryFactory {
     }
   }
 
-  private <T extends MQConnectionFactory> T createConnectionFactoryInstance(Class<T> factoryClass)
-      throws InstantiationException, IllegalAccessException, IllegalArgumentException, InvocationTargetException, NoSuchMethodException, SecurityException {
+  private <T extends MQConnectionFactory> T createConnectionFactoryInstance(Class<T> factoryClass) throws InstantiationException,
+      IllegalAccessException, IllegalArgumentException, InvocationTargetException, NoSuchMethodException, SecurityException {
     logger.trace("createConnectionFactoryInstance for class " + factoryClass.getSimpleName());
     return factoryClass.getConstructor().newInstance();
   }
@@ -378,8 +412,8 @@ public class MQConnectionFactoryFactory {
       catch (Throwable e) {
         logger.warn("Cannot find value of property " + k);
       }
-      logger.trace(String.format("getJndiContext: Using additional property '%s' with value '%s'",k,v));
-      environment.put(k,v);
+      logger.trace(String.format("getJndiContext: Using additional property '%s' with value '%s'", k, v));
+      environment.put(k, v);
     }
 
     ctx = new InitialDirContext(environment);
@@ -395,5 +429,31 @@ public class MQConnectionFactoryFactory {
 
   static boolean isNotNullOrEmpty(String s) {
     return !isNullOrEmpty(s);
+  }
+
+  private SSLSocketFactory getSSLSocketFactory(String b) {
+    SSLSocketFactory sf = null;
+
+    if (b == null || b.isEmpty()) {
+      /* Should never get here as the caller has already checked */
+      logger.trace("getSSLSocketFactory - null/empty bundle name requested");
+      return sf;
+    }
+
+    if (sslBundles != null) {
+      try {
+        SslBundle sb = sslBundles.getBundle(b);
+        SSLContext sc;
+        logger.trace("SSL Bundle for {} - found", b);
+
+        sc = sb.createSslContext();
+        // logger.trace("SSL Protocol is {}",sc.getProtocol());
+        sf = sc.getSocketFactory();
+      }
+      catch (NoSuchSslBundleException e) {
+        logger.error("SSL bundle for {} - not found", b);
+      }
+    }
+    return sf;
   }
 }